Reorganised some stuff
